requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
duckdb==1.2.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.4
python-multipart==0.0.9
python-dateutil==2.9.0
anthropic==0.40.0
langgraph==0.3.5
langchain-anthropic==0.3.0
httpx==0.27.0
torch>=2.0
scikit-learn>=1.3
ruff==0.4.4
pytest==8.2.0
